Autocomplete is done

+++ kord-extensions/src/main/kotlin/com/kotlindiscord/kord/extensions/commands/application/DefaultApplicationCommandRegistry.kt
@@ -18,6 +18,7 @@ import com.kotlindiscord.kord.extensions.commands.application.slash.SlashCommand
 import com.kotlindiscord.kord.extensions.commands.application.user.UserCommand
 import dev.kord.common.entity.Snowflake
 import dev.kord.core.behavior.createApplicationCommands
+import dev.kord.core.event.interaction.AutoCompleteInteractionCreateEvent
 import dev.kord.core.event.interaction.ChatInputCommandInteractionCreateEvent
 import dev.kord.core.event.interaction.MessageCommandInteractionCreateEvent
 import dev.kord.core.event.interaction.UserCommandInteractionCreateEvent
@@ -269,16 +270,18 @@ public open class DefaultApplicationCommandRegistry : ApplicationCommandRegistry
             }
         }
 
-        // Finally, we can remove anything that needs to be removed
-        toRemove.forEach {
-            logger.trace { "Removing ${it.type.name} command: ${it.name}" }
+        if (toAdd.isEmpty() && toUpdate.isEmpty()) {
+            // Finally, we can remove anything that needs to be removed
+            toRemove.forEach {
+                logger.trace { "Removing ${it.type.name} command: ${it.name}" }
 
-            @Suppress("MagicNumber")  // not today, Detekt
-            try {
-                it.delete()
-            } catch (e: KtorRequestException) {
-                if (e.status.code != 404) {
-                    throw e
+                @Suppress("MagicNumber")  // not today, Detekt
+                try {
+                    it.delete()
+                } catch (e: KtorRequestException) {
+                    if (e.status.code != 404) {
+                        throw e
+                    }
                 }
             }
         }
@@ -395,6 +398,35 @@ public open class DefaultApplicationCommandRegistry : ApplicationCommandRegistry
         command ?: return logger.warn { "Received interaction for unknown user command: $commandId" }
 
         command.call(event)
+    }
+
+    override suspend fun handle(event: AutoCompleteInteractionCreateEvent) {
+        val commandId = event.interaction.command.rootId
+        val command = slashCommands[commandId]?.findCommand(event)
+
+        command ?: return logger.warn { "Received autocomplete interaction for unknown command: $commandId" }
+
+        if (command.arguments == null) {
+            return logger.trace { "Command $command doesn't have any arguments." }
+        }
+
+        val option = event.interaction.command.options.filterValues { it.focused }.toList().firstOrNull()
+
+        option ?: return logger.trace { "Autocomplete event for command $command doesn't have a focused option." }
+
+        val arg = command.arguments!!().args.firstOrNull { it.displayName == option.first }
+
+        arg ?: return logger.warn {
+            "Autocomplete event for command $command has an unknown focused option: ${option.first}."
+        }
+
+        val callback = arg.converter.genericBuilder.autoCompleteCallback
+
+        callback ?: return logger.trace {
+            "Autocomplete event for command $command has an focused option without a callback: ${option.first}."
+        }
+
+        callback(event.interaction, event)
     }
 
     // endregion
